Task: cause nose tests to fail if coverage low
In order to maintain quality, the automatic tests should cover all of
the project's code.

Note: Currently there is a bug that causes an error from the cover
plugin for nose when run with both `--cover-branches` and
`--cover-min-percentage`. In the mean time, the automated testing
ignores this error, but once it is fixed everything should work as
expected. For more info see nose-devs/nose/gh-626.